Monitored wrapper immediately starts monitor
