+++ gym/monitoring/tests/test_monitor.py
@@ -4,7 +4,7 @@ import os
 import gym
 from gym import error, spaces
 from gym import monitoring
-from gym import Monitored
+from gym.wrappers import Monitored
 from gym.monitoring import monitor
 from gym.monitoring.tests import helpers
 
@@ -14,45 +14,41 @@ class FakeEnv(gym.Env):
 
 def test_monitor_filename():
     with helpers.tempdir() as temp:
-        env = gym.make('Cartpole-v0')
-        monitor = monitoring.Monitor(env)
-        env = Monitored(env, monitor)
-        monitor.start(temp)
-        monitor.close()
+        env = Monitored(gym.make('CartPole-v0'), temp)
+        env.close()
 
         manifests = glob.glob(os.path.join(temp, '*.manifest.*'))
         assert len(manifests) == 1
 
 def test_write_upon_reset_false():
     with helpers.tempdir() as temp:
-        env = Monitored(gym.make('CartPole-v0'))
-        env.monitor.start(temp, video_callable=False, write_upon_reset=False)
+        env = Monitored(gym.make('CartPole-v0'), 
+            temp, video_callable=False, write_upon_reset=False)
         env.reset()
 
         files = glob.glob(os.path.join(temp, '*'))
         assert not files, "Files: {}".format(files)
 
-        env.monitor.close()
+        env.close()
         files = glob.glob(os.path.join(temp, '*'))
         assert len(files) > 0
 
 def test_write_upon_reset_true():
     with helpers.tempdir() as temp:
-        env = Monitored(gym.make('CartPole-v0'))
-        env.monitor.start(temp, video_callable=False, write_upon_reset=True)
+        env = Monitored(gym.make('CartPole-v0'), 
+            temp, video_callable=False, write_upon_reset=True)
         env.reset()
 
         files = glob.glob(os.path.join(temp, '*'))
         assert len(files) > 0, "Files: {}".format(files)
 
-        env.monitor.close()
+        env.close()
         files = glob.glob(os.path.join(temp, '*'))
         assert len(files) > 0
 
 def test_close_monitor():
     with helpers.tempdir() as temp:
-        env = Monitored(FakeEnv())
-        env.monitor.start(temp)
+        env = Monitored(FakeEnv(), temp)
         env.monitor.close()
 
         manifests = monitor.detect_training_manifests(temp)
@@ -60,9 +56,9 @@ def test_close_monitor():
 
 def test_video_callable_true_not_allowed():
     with helpers.tempdir() as temp:
-        env = Monitored(gym.make('CartPole-v0'))
+        env = gym.make('CartPole-v0')
         try:
-            env.monitor.start(temp, video_callable=True)
+            env = Monitored(env, temp, video_callable=True)
         except error.Error:
             pass
         else:
@@ -70,26 +66,23 @@ def test_video_callable_true_not_allowed():
 
 def test_video_callable_false_does_not_record():
     with helpers.tempdir() as temp:
-        env = Monitored(gym.make('CartPole-v0'))
-        env.monitor.start(temp, video_callable=False)
+        env = Monitored(gym.make('CartPole-v0'), temp, video_callable=False)
         env.reset()
-        env.monitor.close()
+        env.close()
         results = monitoring.load_results(temp)
         assert len(results['videos']) == 0
 
 def test_video_callable_records_videos():
     with helpers.tempdir() as temp:
-        env = Monitored(gym.make('CartPole-v0'))
-        env.monitor.start(temp)
+        env = Monitored(gym.make('CartPole-v0'), temp)
         env.reset()
-        env.monitor.close()
+        env.close()
         results = monitoring.load_results(temp)
         assert len(results['videos']) == 1, "Videos: {}".format(results['videos'])
 
 def test_env_reuse():
     with helpers.tempdir() as temp:
-        env = Monitored(gym.make('CartPole-v0'))
-        env.monitor.start(temp)
+        env = Monitored(gym.make('CartPole-v0'), temp)
         env.monitor.close()
 
         env.monitor.start(temp, force=True)
@@ -121,9 +114,8 @@ gym.envs.register(
 )
 def test_env_reuse():
     with helpers.tempdir() as temp:
-        env = Monitored(gym.make('Autoreset-v0'))
-        env.monitor.start(temp)
-
+        env0 = gym.make('Autoreset-v0')
+        env = Monitored(gym.make('Autoreset-v0'), temp)
         env.reset()
 
         env.step(None)
@@ -145,14 +137,14 @@ def test_no_monitor_reset_unless_done():
 
     with helpers.tempdir() as temp:
         # Make sure we can reset as we please without monitor
-        env = Monitored(gym.make('CartPole-v0'))
+        env = gym.make('CartPole-v0')
         env.reset()
         env.step(env.action_space.sample())
         env.step(env.action_space.sample())
         env.reset()
 
         # can reset once as soon as we start
-        env.monitor.start(temp, video_callable=False)
+        env = Monitored(env, temp, video_callable=False)
         env.reset()
 
         # can reset multiple times in a row
@@ -174,13 +166,12 @@ def test_no_monitor_reset_unless_done():
         env.step(env.action_space.sample())
         assert_reset_raises(env)
 
-        env.monitor.close()
+        env.close()
 
 def test_only_complete_episodes_written():
     with helpers.tempdir() as temp:
-        env = Monitored(gym.make('CartPole-v0'))
+        env = Monitored(gym.make('CartPole-v0'), temp, video_callable=False)
 
-        env.monitor.start(temp, video_callable=False)
         env.reset()
         d = False
         while not d:
@@ -189,7 +180,7 @@ def test_only_complete_episodes_written():
         env.reset()
         env.step(env.action_space.sample())
 
-        env.monitor.close()
+        env.close()
 
         # Only 1 episode should be written
         results = monitoring.load_results(temp)
