+++ gym/monitoring/tests/test_monitor_envs.py
@@ -6,7 +6,7 @@ import logging
 logger = logging.getLogger(__name__)
 
 from gym import envs
-from gym import Monitored
+from gym.wrappers import Monitored
 from gym.monitoring.tests import helpers
 
 specs = [spec for spec in sorted(envs.registry.all(), key=lambda x: x.id) if spec._entry_point is not None]
@@ -26,12 +26,11 @@ def test_renderable_after_monitor_close(spec):
         return
 
     with helpers.tempdir() as temp:
-        env = Monitored(spec.make())
+        env = Monitored(spec.make(), temp)
         # Skip un-renderable envs
         if 'human' not in env.metadata.get('render.modes', []):
             return
 
-        env.monitor.start(temp)
         env.reset()
         env.monitor.close()
