+++ gym/wrappers/__init__.py
@@ -1,1 +1,2 @@
 from gym.wrappers.frame_skipping import SkipWrapper
+from gym.wrappers.monitoring import Monitored
